Refactor: addResponse 추출
