+++ src/page/ChatPage/index.tsx
@@ -1,16 +1,10 @@
-import React, { Dispatch, SetStateAction } from "react";
-import MessageInput from "../../components/MessageInput";
-import CommunicationDisplay from "../../components/CommunicationDisplay";
-import Message from "../../Interface/Message/Message";
+import React from "react";
 import { introMessage } from "../../Data/Message";
-import requestLexResponse from "../../api/LexRequest";
+import Message from "../../Interface/Message/Message";
 import RequestMessageType from "../../Interface/Message/RequestMessageType";
-import {
-	BasicResponseMessageType,
-	ContentResponseMessageType,
-	ResponseMessageType,
-} from "../../Interface/Message/ResponseMessageType";
-import { Grid, Stack } from "@mui/material";
+import Conversation from "../../components/Conversation";
+import MessageInput from "../../components/MessageInput";
+import { addResponse } from "./addResponse";
 
 const createRequest = (message: string) => {
 	return {
@@ -19,30 +13,6 @@ const createRequest = (message: string) => {
 		userId: window.navigator.userAgent,
 	} as RequestMessageType;
 };
-
-async function addResponse(
-	text: string,
-	setMessages: Dispatch<SetStateAction<Message[]>>
-) {
-	const data: ResponseMessageType = {
-		type: "response",
-		content: [] as ContentResponseMessageType[],
-	};
-
-	try {
-		const response = await requestLexResponse(text);
-		data.content = response.data as ContentResponseMessageType[];
-	} catch {
-		data.content = [
-			{
-				contentType: "PlainText",
-				content: "Sorry, I am not able to understand you. Please try again.",
-			},
-		] as ContentResponseMessageType[];
-	}
-
-	setMessages((messages) => [...messages, data]);
-}
 
 export default function ChatPage() {
 	const [messages, setMessages] = React.useState<Message[]>([introMessage]);
@@ -63,19 +33,11 @@ export default function ChatPage() {
 
 	return (
 		<>
-			<Grid
-				container
-				direction="column"
-				flexGrow={"1"}
-				overflow={"scroll"}
-				style={{ WebkitOverflowScrolling: "touch" }}
-			>
-				<CommunicationDisplay
-					loading={loading}
-					messages={messages}
-					onButtonClick={onResponseCardButtonClick}
-				/>
-			</Grid>
+			<Conversation
+				loading={loading}
+				messages={messages}
+				onButtonClick={onResponseCardButtonClick}
+			/>
 			<MessageInput loading={loading} onClick={addMessage} />
 		</>
 	);
